Test out the Box - Maps, LocalStorage e + Code Hacks

+++ qtruck/backend/src/controllers/HelpersController.ts
@@ -1,15 +1,25 @@
 import { Request, Response } from 'express'
 
 import User from '../models/User'
+import FoodTruck from '../models/FoodTruck'
+import Reviews from '../models/Reviews'
 
 export default {
+
     async resetUser(req: Request, res: Response) {
 
-        //recebe um valor de instagram, quando a rota helpers-reset for chamada
         const { instagram } = req.query
-        //para apagar um usuário atravez do valor instagram que foi filtrado na query string com o instagram X
-        await User.deleteMany({ instagram: instagram })
-        //Passando o status 204 tudo certo sem resposta no corpo com end
+
+        const user = await User.findOne({instagram: instagram})
+
+        if (!user)
+            return res.status(204).end()
+
+        await Reviews.deleteMany({rated_by: user._id})
+        await FoodTruck.deleteMany({suggested_by: user._id})
+        await User.deleteMany({instagram: instagram})
+
         return res.status(204).end()
     }
+
 }
